ai_card_generator/requirements.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Web Server Dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Firebase Dependencies
firebase-admin>=6.2.0

# Utility Dependencies
argparse  # Built-in, listed for clarity

# Optional Dependencies for Enhanced Features
Pillow>=10.0.0  # For image processing if needed
urllib3>=2.0.0  # For HTTP requests
certifi>=2023.7.22  # For SSL certificates

# Development Dependencies (optional)
# pytest>=7.4.0  # For testing
# black>=23.0.0   # For code formatting
# flake8>=6.0.0   # For linting
